+++ src/service/CertificateService.ts
@@ -14,25 +14,32 @@
  * limitations under the License.
  */
 
-import { ConfigParams } from './ConfigService';
-import { BootstrapUtils } from './BootstrapUtils';
-import LoggerFactory from '../logger/LoggerFactory';
-import Logger from '../logger/Logger';
-import { LogType } from '../logger';
+import { writeFileSync } from 'fs';
 import { join, resolve } from 'path';
+import { Convert } from 'symbol-sdk';
+import { LogType } from '../logger';
+import Logger from '../logger/Logger';
+import LoggerFactory from '../logger/LoggerFactory';
 import { CertificatePair, ConfigPreset } from '../model';
+import { BootstrapUtils } from './BootstrapUtils';
+import { ConfigParams } from './ConfigService';
 
 type CertificateParams = ConfigParams;
 
 const logger: Logger = LoggerFactory.getLogger(LogType.System);
 
 interface NodeCertificates {
-    ssl: CertificatePair;
-    node: CertificatePair;
+    main: CertificatePair;
+    transport: CertificatePair;
 }
 
 export class CertificateService {
     constructor(private readonly root: string, protected readonly params: CertificateParams) {}
+
+    public static toAns1(privateKey: string): string {
+        const prefix = '302e020100300506032b657004220420';
+        return `${prefix}${privateKey.toLowerCase()}`;
+    }
 
     public static getCertificates(stdout: string): CertificatePair[] {
         const locations = (string: string, substring: string): number[] => {
@@ -67,31 +74,35 @@ export class CertificateService {
         });
     }
 
-    public async run(name: string): Promise<NodeCertificates> {
-        // Currently the SSL certifcates are created via a docker image. Migrate this to a native forge!
-        // https://www.npmjs.com/package/node-forge
-
-        const presetData: ConfigPreset = BootstrapUtils.loadPresetData(
-            this.root,
-            this.params.preset,
-            this.params.assembly,
-            this.params.customPreset,
-            this.params.customPresetObject,
-        );
-        const symbolServerToolsImage = presetData.symbolServerToolsImage;
+    public async run(presetData: ConfigPreset, name: string, providedCertificates: NodeCertificates): Promise<void> {
         const copyFrom = `${this.root}/config/cert`;
         const certFolder = BootstrapUtils.getTargetNodesFolder(this.params.target, false, name, 'userconfig', 'resources', 'cert');
         await BootstrapUtils.mkdir(certFolder);
-        await BootstrapUtils.mkdir(join(certFolder, 'new_certs'));
+        const newCertsFolder = join(certFolder, 'new_certs');
+        await BootstrapUtils.mkdir(newCertsFolder);
         const generatedContext = { name };
-        const templateContext: any = { ...presetData, ...generatedContext };
-        await BootstrapUtils.generateConfiguration(templateContext, copyFrom, certFolder);
+        await BootstrapUtils.generateConfiguration(generatedContext, copyFrom, certFolder, []);
 
+        // if (!BootstrapUtils.isWindows()) {
+        //     // chmodSync(newCertsFolder, 700);
+        // }
+        // await BootstrapUtils.writeTextFile(join(certFolder, 'index.txt'), '');
+        // await BootstrapUtils.writeTextFile(join(certFolder, 'index.txt.attr'), '');
+        // await BootstrapUtils.writeTextFile(join(certFolder, 'serial.dat'), Convert.uint8ToHex(Crypto.randomBytes(19)).toUpperCase());
+
+        writeFileSync(join(certFolder, 'ca.der'), Convert.hexToUint8(CertificateService.toAns1(providedCertificates.main.privateKey)));
+        writeFileSync(
+            join(certFolder, 'node.der'),
+            Convert.hexToUint8(CertificateService.toAns1(providedCertificates.transport.privateKey)),
+        );
+
+        // TODO. Migrate this process to forge, sshpk or any node native implementation.
         const command = this.createCertCommands('/data');
         await BootstrapUtils.writeTextFile(join(certFolder, 'createCertificate.sh'), command);
         const cmd = ['bash', 'createCertificate.sh'];
         const binds = [`${resolve(certFolder)}:/data:rw`];
         const userId = await BootstrapUtils.resolveDockerUserFromParam(this.params.user);
+        const symbolServerToolsImage = presetData.symbolServerToolsImage;
         const { stdout, stderr } = await BootstrapUtils.runImageUsingExec({
             catapultAppFolder: presetData.catapultAppFolder,
             image: symbolServerToolsImage,
@@ -111,19 +122,26 @@ export class CertificateService {
             throw new Error('Certificate creation failed. 2 certificates should have been created but got: ' + certificates.length);
         }
         logger.info(`Certificate for node ${name} created`);
-        return { ssl: certificates[0], node: certificates[1] };
+        const caCertificate = certificates[0];
+        const nodeCertificate = certificates[1];
+
+        BootstrapUtils.validateIsTrue(caCertificate.privateKey === providedCertificates.main.privateKey, 'Invalid ca private key');
+        BootstrapUtils.validateIsTrue(caCertificate.publicKey === providedCertificates.main.publicKey, 'Invalid ca public key');
+
+        BootstrapUtils.validateIsTrue(nodeCertificate.privateKey === providedCertificates.transport.privateKey, 'Invalid Node private key');
+
+        BootstrapUtils.validateIsTrue(nodeCertificate.publicKey === providedCertificates.transport.publicKey, 'Invalid Node public key');
     }
 
     private createCertCommands(target: string): string {
-        return `
+        return `set -e
 cd ${target}
-set -e
 chmod 700 new_certs
 touch index.txt.attr
 touch index.txt
 
 # create CA key
-openssl genpkey -out ca.key.pem -outform PEM -algorithm ed25519
+cat ca.der | openssl pkey -inform DER -outform PEM -out ca.key.pem
 openssl pkey -inform pem -in ca.key.pem -text -noout
 openssl pkey -in ca.key.pem -pubout -out ca.pubkey.pem
 
@@ -132,7 +150,7 @@ openssl req -config ca.cnf -keyform PEM -key ca.key.pem -new -x509 -days 7300 -o
 openssl x509 -in ca.cert.pem  -text -noout
 
 # create node key
-openssl genpkey -out node.key.pem -outform PEM -algorithm ed25519
+cat node.der | openssl pkey -inform DER -outform PEM -out node.key.pem
 openssl pkey -inform pem -in node.key.pem -text -noout
 
 # create request
@@ -150,6 +168,12 @@ openssl verify -CAfile ca.cert.pem node.crt.pem
 
 # finally create full crt
 cat node.crt.pem ca.cert.pem > node.full.crt.pem
+
+rm createCertificate.sh
+rm ca.key.pem
+rm ca.der
+rm node.der
+
 echo "Certificate Created"
 `;
     }